requirements.txt
Django==4.1
djangorestframework==3.12.4
python-decouple==3.6
psycopg2-binary==2.9.5
pydub
pytz
ffmpeg
ffprobe
PyJWT==2.8.0
boto3